requirements.txt
torch
sentence-transformers
fastapi
uvicorn[standard]
plotly
umap-learn
scikit-learn
pandas
numpy
jinja2
pytesseract
pdf2image
pillow
python-multipart
